+++ eslint.config.mjs
@@ -1,5 +1,5 @@
-import typescriptEslint from "@typescript-eslint/eslint-plugin";
-import tsParser from "@typescript-eslint/parser";
+import typescriptEslint from "@typescript-eslint/eslint-plugin"
+import tsParser from "@typescript-eslint/parser"
 
 export default [{
     files: ["**/*.ts"],
@@ -20,7 +20,6 @@ export default [{
             format: ["camelCase", "PascalCase"],
         }],
 
-        eqeqeq: "warn",
         "no-throw-literal": "warn"
     },
-}];
+}]