+++ out/commands.js
@@ -33,16 +33,59 @@ var __importStar = (this && this.__importStar) || (function () {
     };
 })();
 Object.defineProperty(exports, "__esModule", { value: true });
+exports.registerDefaultCommands = registerDefaultCommands;
 exports.registerCommands = registerCommands;
 const _1 = require(".");
 const vscode = __importStar(require("vscode"));
+/**
+ * Registers the default commands which do not require extension enablement.
+ * @param ctx The extension context.
+ */
+function registerDefaultCommands(ctx) {
+    ctx.subscriptions.push(
+    // Open Extension Log
+    vscode.commands.registerCommand("forgevsc.openExtensionLog", () => {
+        _1.Logger.show();
+    }), 
+    // Open Settings (Backend)
+    vscode.commands.registerCommand("forgevsc.openSettings", async (setting, user = true) => {
+        await vscode.commands.executeCommand("workbench.action." + (user ? "openSettings" : "openWorkspaceSettings"), setting?.trim() || "@ext:tryforge.forgevsc");
+    }), 
+    // Open Extension Settings (UI)
+    vscode.commands.registerCommand("forgevsc.openExtensionSettings", async () => {
+        const items = [];
+        if (!!vscode.workspace.workspaceFolders?.length) {
+            items.push({
+                label: "$(folder) Workspace Settings",
+                detail: vscode.workspace.name || "Workspace",
+                description: "Folder",
+                target: "workbench.action.openWorkspaceSettings"
+            });
+        }
+        items.push({
+            label: "$(account) User Settings",
+            description: "Global",
+            target: "workbench.action.openSettings"
+        });
+        const choice = await vscode.window.showQuickPick(items, {
+            placeHolder: "Which extension settings would you like to open?"
+        });
+        if (!choice)
+            return;
+        await vscode.commands.executeCommand(choice.target, "@ext:tryforge.forgevsc");
+    }));
+}
+/**
+ * Registers the commands which require extension enablement.
+ * @param ctx The extension context.
+ */
 function registerCommands(ctx) {
     ctx.subscriptions.push(
     // Create Config
     vscode.commands.registerCommand("forgevsc.createConfig", async () => {
         const action = await vscode.window.showWarningMessage("The custom configuration file is deprecated and maintained only for legacy compatibility. Please use extension settings instead.", "Open Settings", "Dismiss");
         if (action === "Open Settings") {
-            await vscode.commands.executeCommand("forgevsc.openExtensionSettings");
+            await vscode.commands.executeCommand("forgevsc.openSettings", undefined, false);
             return;
         }
         const folders = vscode.workspace.workspaceFolders;
@@ -106,10 +149,6 @@ function registerCommands(ctx) {
     vscode.commands.registerCommand("forgevsc.reloadFunctionMetadata", async () => {
         await (0, _1.getFunctions)(true);
         vscode.window.showInformationMessage("Successfully fetched function metadata!");
-    }), 
-    // Open Extension Log
-    vscode.commands.registerCommand("forgevsc.openExtensionLog", async () => {
-        _1.Logger.show();
     }), 
     // Create Guide
     vscode.commands.registerCommand("forgevsc.createGuide", async () => {