+++ out/guides.js
@@ -345,8 +345,8 @@ async function searchGuides() {
                 ? toTitleCase(guide.targetType)
                 : [guide.category, guide.subCategory].filter(Boolean).join(" • "),
             iconPath: new vscode.ThemeIcon("book"),
-            guide,
-            alwaysShow: true
+            alwaysShow: true,
+            guide
         }));
         const usedKeys = Object.keys(filters);
         const keyItems = keys
@@ -355,7 +355,8 @@ async function searchGuides() {
             label: key + ":",
             description: "Filter",
             action: "key",
-            alwaysShow: true
+            alwaysShow: true,
+            iconPath: new vscode.ThemeIcon("filter")
         }));
         let suggestionItems = [];
         if (!!activeKey && activeKey && values[activeKey]) {
@@ -367,8 +368,8 @@ async function searchGuides() {
                 description: `${activeKey}`,
                 action: "set-filter",
                 key: activeKey,
-                value,
-                alwaysShow: true
+                alwaysShow: true,
+                value
             }));
         }
         qp.items = [